added support for (possible) future Bacteria
changed queen bee mating logic, changed some of the output so it looks better. also, there should not be any bugs at this point regarding the counting of bees or food, be it collected or up for grabs.
bees have synchronized clocks now, a day is a day for all of them, and other small quality of life additions.